distrubute the schema around in a more modular manner
