+++ server/schema.graphql
@@ -1,13 +1,13 @@
 type User {
   id: Int!
-  bio String
+  bio: String
   chirps: [Chirp]
   follows: [Follow]
   likes: [Like]
-  password String!
-  pinnedChirp Int
-  privacyStatus Int!
-  profilePicture String
+  password: String
+  pinnedChirp: Int
+  privacyStatus: Int
+  profilePicture: String
   username: String!
 }
 
@@ -25,4 +25,9 @@ type Like{
 type Follow{
   follower_id: Users.id!
   followee_id: Users.id!
+}
+## testing
+type Book {
+  title: String
+  author: String
 }
