+++ server/schema.js
@@ -1,0 +1,36 @@
+const { gql } = require('apollo-server');
+
+const typeDefs = gql`
+  type User {
+    id: Int!
+    bio: String
+    password: String
+    pinnedChirp: Int
+    privacyStatus: Int
+    profilePicture: String
+    username: String!
+  }
+
+  type Chirp {
+    id: Int!
+    content: String!
+    user_id: Int!
+  }
+
+  type Book {
+    title: String
+    author: String
+  }
+
+  type Query {
+    books: [Book]
+    chirps: [Chirp]
+    users: [User]
+  }
+
+  type Mutation {
+    createUser(id: ID, username: String): String
+  }
+`;
+
+module.exports = typeDefs;
